requirements.txt
langchain==0.3.3
langchain-core==0.3.13
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-groq==0.2.0
pinecone-client==4.0.0
python-dotenv